+++ src/core/settings.py
@@ -40,6 +40,7 @@ DEBUG = env("DEBUG")
 SECRET_KEY = env("SECRET_KEY")
 ALLOWED_HOSTS = env("ALLOWED_HOSTS").split(" ")
 MAINTENANCE_MODE = env("MAINTENANCE_MODE")
+PREPEND_WWW = env.bool("PREPEND_WWW", default=False)
 CSRF_TRUSTED_ORIGINS = [f"https://{allowed_host}" for allowed_host in ALLOWED_HOSTS]
 
 DATABASES = {